requirements.txt
langgraph
langchain-groq
langchain-mcp-adapters
duckdb
fastapi
uvicorn
streamlit
pandas
python-dotenv
mcp
fastmcp
requests
